[adapters/memcache] swallow RingError exception
Currently, if memcache is not available peek will 500 your app.
I believe peek is not *essential* functionality and should not bring
down your app. I would like to print an error, though currently there is
no defined pattern of where to print and I'm not ready to $STDERR.puts
yet.