Updated README to include Flexirest Peek
Flexirest is for accessing REST services in an ActiveRecord style. ActiveResource already exists for this, but it doesn't work where the resource naming doesn't follow Rails conventions, it doesn't have in-built caching and it's not as flexible in general.  It's a supported fork of ActiveRestClient.

I've just created a Peek gem for it.