Handle concurrent requests
When using Peek with a multithreaded application server (like Puma),
Peek.request_id was a problem: it was itself an atomic reference, but it
could be updated by multiple requests before they needed to read it. The
flow was:

1. Requests starts.
2. Set Peek.request_id from Rails' request ID.
3. Perform request.
4. Save results using Peek.request_id.
5. Clear Peek.request_id.

Another thread could do steps 2 or 5 for a different request before step
4 happened for 'this' request.

As Peek uses ActiveSupport instrumentation, and the
`process_action.action_controller` event has access to the request
details, we actually didn't need Peek.request_id at all: we can pass the
ID from the request to the adapter's `#save` method directly.