Do not persist request data in Rails 4.2
See: https://github.com/peek/peek/issues/116

In Rails 4.2, some notification payloads do not include headers.

I'm not entirely sure how to run the tests against the 4.2 Gemfile, but
I was able to figure out how to bypass the exception gently.

If someone wants to point me towards how to actually solve the problem,
I'd be down to try it!